finished dashboard refactor with all functionality in tact
